Corrected User model + improved folder structure
